src/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==25.3.0
billiard==4.2.4
celery==5.6.3
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20241021
distro==1.9.0
Django==5.2.13
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==24.2
django-import-export==4.4.0
django-jazzmin==3.0.4
django-js-asset==3.1.2
django-modeltranslation==0.18.11
djangorestframework==3.17.1
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-yasg==1.21.10
fastapi==0.135.3
groq==1.1.2
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.6.2
packaging==26.0
pillow==12.2.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
PyJWT==2.10.1
pytest==9.0.3
python-dateutil==2.8.2
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2024.1
PyYAML==6.0.2
redis==7.4.0
referencing==0.36.2
requests==2.31.0
rpds-py==0.30.0
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
some-package==0.1
sqlparse==0.5.5
starlette==1.0.0
tablib==3.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.6.3
uvicorn==0.44.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
